third_party/WebKit/LayoutTests/fast/compositorworker/compositor-attribute-change.html

<!DOCTYPE html>
<div id="container"></div>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>

test(function() {
    var proxy = new CompositorProxy(document.getElementById('container'), ['opacity']);
    assert_throws("NoModificationAllowedError", function() { proxy.opacity = 0.4; });
    assert_throws("NoModificationAllowedError", function() { return proxy.opacity; });
}, "This test checks that a compositor proxy cannot be mutated from the main thread.");

</script>
